requirements.txt
torch
torchvision
opencv-python
Pillow
pylance
scikit-learn
pandas
numpy
matplotlib
streamlit
ultralytics
tensorboard
pytest
tqdm